requirements.txt
clip @ git+https://github.com/openai/CLIP.git@40f5484c1c74edd83cb9cf687c6ab92b28d8b656
ftfy==6.1.1
idna==3.3
numpy==1.22.0
pandas==1.3.5
Pillow==9.1.1
python-dateutil==2.8.2
pytz==2022.1
regex==2022.3.15
requests==2.27.1
scipy==1.7.3
six==1.16.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
PyYaml==6.0
easydict==1.10
keytotext==2.3.2
opencv-python==4.7.0.68
scikit-image==0.19.3